Extract links from markdown documents
Proof of concept for extracting basic markdown links from documents.
This isn't necessary for converting markdown to HTML but this data could
be used to provide a list of related notes when rendering the current
one.

I want to add this now because it will make it easier to introduce
wikilink syntax, e.g. [[link]]

Wikilink syntax is one of the few pieces of syntax Obsidian supports that
Github-Flavoured Markdown doesn't (see
https://help.obsidian.md/Editing+and+formatting/Obsidian+Flavored+Markdown)
and supporting it is essential for generating useful HTML.